+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Updates in wallet preferences dialog
 - Update breez to 0.5.1-rc5
+- Improve lightning account deletion
 
 ## [4.0.30] - 2024-06-13